+++ test/suggested_indexes_test.rb
@@ -24,4 +24,16 @@ class SuggestedIndexesTest < Minitest::Test
     User.where("updated_at > ?", Time.now).to_a
     assert_equal [], database.suggested_indexes.map { |q| q.except(:queries, :details) }
   end
+
+  def test_gist_trgm
+    query = "SELECT * FROM users WHERE country = 'Test 1'"
+    result = database.suggested_indexes_by_query(queries: [query])[query]
+    assert_nil result[:covering_index]
+  end
+
+  def test_ltree
+    query = "SELECT * FROM users WHERE tree_path = 'path1'"
+    result = database.suggested_indexes_by_query(queries: [query])[query]
+    assert_equal ["tree_path"], result[:covering_index]
+  end
 end